Move ascend utility files to sglang.test.ascend
Move gsm8k_ascend_mixin.py and vlm_utils.py (renamed from test_vlm_utils.py)
out of registered/ since they are utility classes, not tests with CI registration.
This fixes the sanity check error.
