+++ python/sglang/test/ascend/vlm_utils.py
