+++ test/registered/ascend/llm_models/test_ascend_persimmon_8b_chat.py
@@ -1,6 +1,6 @@
 import unittest
 
-from gsm8k_ascend_mixin import GSM8KAscendMixin
+from sglang.test.ascend.gsm8k_ascend_mixin import GSM8KAscendMixin
 
 from sglang.test.ci.ci_register import register_npu_ci
 from sglang.test.test_utils import CustomTestCase
